drf-ecom/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
Pillow==10.1.0
django-filter==23.5
drf-spectacular==0.27.0
python-decouple==3.8